phong push get product home

+++ app/src/main/java/com/example/duantotnghiep_md27/Model/ProductData.java
@@ -28,7 +28,7 @@ public class ProductData {
         this.productList = productList;
     }
 
-    public int getStatus() {
+    public double getStatus() {
         return status;
     }
